feat: move schemas out of lib
